fix auto router description textinput
